requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
jinja2
weasyprint
python-dotenv
openai
